duplicate debugging removed
Expr.expr already returns what was printed by this debugging statement,
so there's no reason to print it as it comes in... the goal of the app
is to print the eval result of Expr.expr; maybe Expr.expr(array).eval()
returns an int or a string.

We've learned to do this in many languages using a token classification
method, with Term, Op: + -, Prod: * / %, and ( Term ), where Term also
parses out as Term Op Term, and Prod is an instance of Term as well.